D09P01 - Solved

+++ src/main/kotlin/be/inniger/advent/Day02.kt
@@ -8,14 +8,14 @@ class Day02 {
         private const val DESIRED_OUTPUT = 19690720
     }
 
-    fun solveFirst(program: List<Int>, noun: Int = program[1], verb: Int = program[2]) =
+    fun solveFirst(program: List<Long>, noun: Long = program[1], verb: Long = program[2]) =
         IntComputer(editNounAndVerb(program, Words(noun, verb))).runProgram().firsPositionValue
 
-    fun solveSecond(program: List<Int>) = generateAllWords()
+    fun solveSecond(program: List<Long>) = generateAllWords()
         .first { IntComputer(editNounAndVerb(program, it)).runProgram().firsPositionValue.toInt() == DESIRED_OUTPUT }
         .let { 100 * it.noun + it.verb }
 
-    private fun editNounAndVerb(program: List<Int>, words: Words): List<Int> {
+    private fun editNounAndVerb(program: List<Long>, words: Words): List<Long> {
         val mutableProgram = program.toMutableList()
 
         mutableProgram[1] = words.noun
@@ -26,8 +26,8 @@ class Day02 {
 
     private fun generateAllWords() =
         (1..99).flatMap { noun ->
-            (1..99).map { verb -> Words(noun, verb) }
+            (1..99).map { verb -> Words(noun.toLong(), verb.toLong()) }
         }
 
-    private data class Words(val noun: Int, val verb: Int)
+    private data class Words(val noun: Long, val verb: Long)
 }
